--- src/main/java/com/pagely/aiservice/ai/domain/UserLikedBook.java
@@ -1,43 +1,0 @@
-package com.pagely.aiservice.ai.domain;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.Id;
-import jakarta.persistence.PrePersist;
-import jakarta.persistence.Table;
-import java.time.LocalDateTime;
-import java.util.UUID;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.hibernate.annotations.UuidGenerator;
-
-@Entity
-@Table(name = "p_user_liked_book")
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-@Builder
-public class UserLikedBook {
-
-    @Id
-    @UuidGenerator
-    @Column(nullable = false, updatable = false)
-    private UUID id;
-
-    @Column(nullable = false)
-    private UUID userId;
-
-    @Column(nullable = false)
-    private String bookId;
-
-    // 도서 정보 반정규화
-    private String title;
-    private String author;
-    private String category;
-
-    @Column(columnDefinition = "TEXT")
-    private String description;
-}
