--- src/main/java/com/pagely/aiservice/ai/domain/UserSearchHistory.java
@@ -1,51 +1,0 @@
-package com.pagely.aiservice.ai.domain;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.Id;
-import jakarta.persistence.PrePersist;
-import jakarta.persistence.Table;
-import java.time.LocalDateTime;
-import java.util.UUID;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.hibernate.annotations.UuidGenerator;
-
-@Entity
-@Table(name = "p_user_search_history")
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-@Builder
-public class UserSearchHistory {
-
-    @Id
-    @UuidGenerator
-    @Column(nullable = false, updatable = false)
-    private UUID id;
-
-    @Column(nullable = false)
-    private UUID userId;
-
-    // 검색 후 클릭한 도서 (검색만 하고 클릭 안 했을 수도 있으므로 nullable)
-    private String bookId;
-
-    // 도서 정보 반정규화
-    private String title;
-    private String author;
-    private String category;
-
-    private String searchKeyword;
-
-    private LocalDateTime searchedAt;
-
-    @PrePersist
-    protected void prePersist() {
-        if (searchedAt == null) {
-            searchedAt = LocalDateTime.now();
-        }
-    }
-}
