+++ src/main/java/com/pagely/aiservice/ai/infrastructure/persistence/UserLikedBookJpaRepository.java
@@ -1,6 +1,6 @@
 package com.pagely.aiservice.ai.infrastructure.persistence;
 
-import com.pagely.aiservice.ai.domain.UserLikedBook;
+import com.pagely.aiservice.ai.domain.model.UserLikedBook;
 import java.util.List;
 import java.util.UUID;
 import org.springframework.data.jpa.repository.JpaRepository;
