+++ src/main/java/com/pagely/aiservice/ai/infrastructure/persistence/UserSearchHistoryJpaRepository.java
@@ -1,6 +1,6 @@
 package com.pagely.aiservice.ai.infrastructure.persistence;
 
-import com.pagely.aiservice.ai.domain.UserSearchHistory;
+import com.pagely.aiservice.ai.domain.model.UserSearchHistory;
 import java.util.List;
 import java.util.UUID;
 import org.springframework.data.jpa.repository.JpaRepository;
